FramePerSec double,  not int
git-svn-id: https://svn.atnf.csiro.au/difx@8293 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/codifio/trunk/src/codifio.c
@@ -104,7 +104,7 @@ int getCODIFFrameMJD(const codif_header *header)
   return mjd + header->seconds/86400; // Seconds will usually be greater than one day
 }
 
-double getCODIFFrameDMJD(const codif_header *header, int framepersec) 
+double getCODIFFrameDMJD(const codif_header *header, double framepersec) 
 {
   int mjd = getCODIFFrameMJD(header);
   int sec = getCODIFFrameSecond(header);
